Proyectos/Proyecto 2/Proyecto2_IA-main/requirements.txt
flask
opencv-python
pillow
transformers
numpy
gTTS
easyocr
language-tool-python
edge-tts

